requirements.txt
streamlit
pandas
fpdf
pillow
openpyxl